### 2020-07-10 #2470 @ritagnunes
The full URL should be:  https://github.com/ritagnunes/PulseqDiffusion
Thanks
