### 2020-07-10 #2470 @kyleniemeyer
@arfon can we tell whedon to update the repo URL to https://github.com/ritagnunes/PulseqDiffusion for checking and building the paper?
