### 2020-07-10 #2470 @arfon
> @arfon can we tell whedon to update the repo URL to https://github.com/ritagnunes/PulseqDiffusion for checking and building the paper?

I've fixed this on the back end too. For now there's not a Whedon-way to do this.
