### 2020-07-12 #2470 @Kevin-Mattheus-Moerman
@grlee77, @mathieuboudreau, @spinicist thanks for your review efforts to date! Would you be interested in helping with this review? It is related to this work you reviewed for us: https://github.com/openjournals/joss-reviews/issues/1725
@mfroeling this also looks like your cup of tea, would you be interested in helping? Thanks!

Paper: https://github.com/openjournals/joss-papers/blob/joss.02470/joss.02470/10.21105.joss.02470.pdf
Software: https://github.com/ritagnunes/PulseqDiffusion

Also do let me know if I am overloading you with review requests.
