### 2020-07-12 #2470 @Kevin-Mattheus-Moerman
> I can, but I'm only available in about a week as I'm currently on vacation.

That is fine thank you so much!
