### 2020-07-12 #2478 @whedon
[REVIEW]: PulseqDiffusion: PyPulseq implementation of Diffusion-Weighted Echo Planar Imaging
**Submitting author:** @ritagnunes (<a href="http://orcid.org/0000-0001-7425-5717">Rita G. Nunes</a>)
**Repository:** <a href="https://github.com/ritagnunes/PulseqDiffusion" target ="_blank">https://github.com/ritagnunes/PulseqDiffusion</a>
**Version:** v1.0.0
