### 2020-07-12 4ca5652
Update paper.bib
Fixed doi issues
