+++ paper/paper.bib
@@ -39,7 +39,7 @@
   number = {2},
   pages = {782--790},
   year = {2012},
-  doi = {https://doi.org/10.1016/j.neuroimage.2011.09.015},
+  doi = {10.1016/j.neuroimage.2011.09.015},
   keywords = {FSL, Software},
 }
 
@@ -102,7 +102,7 @@
 
 @article{ravi2019pypulseqjoss,
   doi = {10.21105/joss.01725},
-  url = {https://doi.org/10.21105/joss.01725},
+  url = {10.21105/joss.01725},
   year = {2019},
   publisher = {The Open Journal},
   volume = {4},
@@ -145,7 +145,7 @@
   volume = {202},
   pages = {116137},
   year = {2019},
-  doi = {https://doi.org/10.1016/j.neuroimage.2019.116137},
+  doi = {10.1016/j.neuroimage.2019.116137},
   author = {Tournier, J-Donald and Smith, Robert and Raffelt, David and Tabbara, Rami and Dhollander, Thijs and Pietsch, Maximilian and Christiaens, Daan and Jeurissen, Ben and Yeh, Chun-Hung and Connelly, Allan},
   keywords = {MRI, Software, Image, Processing, Visualisation},
 }
