### 2020-07-12 #2478 @Kevin-Mattheus-Moerman
@ritagnunes can you fix those DOI's :point_up:, FYI you can run `@whedon generate pdf` here to update the paper, or run `@whedon check references` to check those DOI's. Thanks
