### 2020-07-12 #2478 @ritagnunes
On Sun, 12 Jul 2020, 22:26 Kevin Mattheus Moerman, <notifications@github.com>
wrote:

> @mathieuboudreau <https://github.com/mathieuboudreau>, @grlee77
> <https://github.com/grlee77> thanks for your help here! I look forward to
> seeing your review comments. I would like to point out that judging from
> the number of lines of code and the content of the repo this submission
> could be on the skinny side of what is typically accepted in JOSS. If you
> feel this submission is too minor (in terms of functionality, research
> application, etc) please do flag this with me. Thanks!
>
