### 2020-07-13 #2478 @grlee77
> @mathieuboudreau, @grlee77 thanks for your help here! I look forward to seeing your review comments. I would like to point out that judging from the number of lines of code and the content of the repo this submission could be on the skinny side of what is typically accepted in JOSS. If you feel this submission is too minor (in terms of functionality, research application, etc) please do flag this with me. Thanks!

I have not done a real review yet, but based on a quick look, the primary contribution is the three files named `write_*.py` [in this folder](https://github.com/ritagnunes/PulseqDiffusion/tree/master/PulseqDiffusion). These files generate three different variants of a vendor-neutral diffusion MRI sequence using the PyPulseq tool I previously helped review for JOSS. There was already a basic EPI example in PyPulseq itself, so the new part is the addition of the diffusion-encoding gradients and either one or two 180 degree refocusing pulses. These turn the pulse sequence from a basic gradient-echo EPI sequence into spin-echo one that is sensitive to diffusion. This type of sequence is widely used in diffusion-based neuroimaging studies.

I am sympathetic to the goal of providing a vendor-neutral implementation of these sequences and am not aware of another existing source for such sequences. In practice, research sites that have a commercial license for a vendor-specific implementation are going to continue to use that for real-world studies, as they are more widely tested and often incorporate additional features (simultaneous multi-slice, parallel imaging, etc.). There is still value in having open source implementations as a teaching tool, at sites without an equivalent commercial sequence, and potentially for some more specialized research cases.

Aside from the pulse sequence generating functions, the authors provide the file `recon_imgs.m` which the user would use to make images from the k-space data acquired by these sequences. The provided reconstruction script looks like it would work for two of the variants, but I'm not sure about the "ramp-sampling" variant. I did not see any code to regrid the non-uniformly spaced samples (due to sampling on the ramps) back to a uniformly spaced grid, but it is possible I missed it somewhere.

The recon script included is pretty basic compared to more general open source MRI reconstruction tools, but it does match the provided pulse sequence and should simplify reconstruction for end users that may not be image reconstruction specialists. Therefore, it does add value above just providing the sequence alone.

In summary, I think the submission has value, but agree that it is a somewhat incremental addition on top of PyPulseq itself.
