### 2020-07-31 ritagnunes/PulseqDiffusion#4 @mathieuboudreau
[JOSS] Installation instructions
Hi,

The installation instructions on the README are somewhate lacking for me to be able to get up and running to use the software.  https://github.com/ritagnunes/PulseqDiffusion/blob/master/README.md

* I have MATLAB installed.
* I created a venv with python 3.6. I downloaded this github repository (even though those aren't in the installation instructions). I installed `pip install pypulseq` in it after activating the venv.
* I installed the NIfTI reader you recommended in Ref 7 of your README and placed it in a folder called 'nifti_tools'.
* I have FSL installed.

Now what am I supposed to do? Is it installed correctly now that I've done those steps?
