### 2020-07-31 ritagnunes/PulseqDiffusion#6 @mathieuboudreau
[JOSS] Functionality documentation missing
Hi,

Per the JOSS guidelines, your should provide the following:

>**Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I can't seem to find anywhere (readme, wiki, readthedocs, etc), where the core functionality of your software is documented. Am I not looking in the right place?
