### 2020-08-03 8ee6332
Update LongSS_Signal.m
Response to: [JOSS] suggested fix to function name/terminology #7
