+++ PulseqDiffusion/SNR_CNR_Study/LongSS_Signal.m
@@ -1,5 +1,5 @@
-function [SS_Signal] = LongSS_Signal(FlipAngle,TRep,TE,T1,T2)
-% Longitudinal Steady-State signal calculation given the flip angle in degrees
+function [Transv_SS_Signal] = TransvSS_Signal(FlipAngle,TRep,TE,T1,T2)
+% Transverse Steady-State signal calculation given the flip angle in degrees
 % by T.T. Fernandes, September 2019 - LarSys - Instituto Superior Tecnico - Universidade de Lisboa
 % usage:SS_Signal = LongSS_Signal(FlipAngle,TRep,TE,T1,T2)
 %
@@ -9,8 +9,8 @@ function [SS_Signal] = LongSS_Signal(FlipAngle,TRep,TE,T1,T2)
 % :parameters: T1: T1 time in seconds
 % :parameters: T2: T2 time in seconds
 %
-% :returns: SS_Signal: Longitudinal Steady-State signal
+% :returns: Transv_SS_Signal: Transverse Steady-State signal
 
-SS_Signal = sind(FlipAngle).*exp(-TE/T2).*(1-exp(-TRep/T1))./(1-(exp(-TRep/T1)).*cosd(FlipAngle));
+Transv_SS_Signal = sind(FlipAngle).*exp(-TE/T2).*(1-exp(-TRep/T1))./(1-(exp(-TRep/T1)).*cosd(FlipAngle));
 
 end
