### 2020-08-03 bd906ad
Update and rename LongSS_SignalTest.m to TransvSS_SignalTest.m
Response to: [JOSS] suggested fix to function name/terminology #7
