### 2020-08-03 37f0dcd
Update READ_SNR_CNR_Study.txt
Response to:  [JOSS] suggested fix to function name/terminology #7
