+++ PulseqDiffusion/SNR_CNR_Study/READ_SNR_CNR_Study.txt
@@ -8,7 +8,7 @@ This is a Script that predicts the signal-to-noise ratio (SNR) observed for
 different brain tissues (gray matter, white matter, cerebrospinal fluid - CSF) and the contrast-to-noise ratio (CNR) of an acute stroke
 lesion relative to those tissues. The implemented code predicts the 
 SNR and CNR per time unit for a spin-echo diffusion-weighted sequence, 
-taking into account the steady-state value for the longitudinal
+taking into account the steady-state value for the transverse
 magnetization. For that purpose, we adapted an expression used to
 predict the SNR per tissue per time unit for a spoiled gradient echo
 sequence (Marques, JP et al, 2019). This simulation considers the possibility to use either EPI or spiral readouts and considers the impact of: B0, max gradient amplitude, spatial resolution and b-value (comparing the achieved SNR with that of a typical 1.5T clinical scanner).
