### 2020-08-03 4c94530
Update SNR_CNR_Study_Simulation.m
Response to:  [JOSS] suggested fix to function name/terminology #7
