+++ PulseqDiffusion/SNR_CNR_Study/SNR_CNR_Study_Simulation.m
@@ -12,7 +12,7 @@
 % fluid - CSF) and the contrast-to-noise ratio (CNR) of an acute stroke
 % lesion relative to those tissues. The implemented code predicts the 
 % SNR and CNR per time unit for a spin-echo diffusion-weighted sequence, 
-% taking into account the steady-state value for the longitudinal
+% taking into account the steady-state value for the transverse
 % magnetization. For that purpose, we adapted an expression used to
 % predict the SNR per tissue per time unit for a spoiled gradient echo
 % sequence (Marques, JP et al, 2019). This simulation considers the possibility to use either
@@ -164,22 +164,22 @@ OptimumAlpha_LS_B0ref = Ernstangle_d(TR_B0ref,T1_LS(refB0idx));
 
 % References values for each tissue (GM, WM, LS) and its respective SNR
 % ... GM tissue ...
-LongSSs_GM_B0ref = LongSS_Signal(OptimumAlpha_GM_B0ref,TR_B0ref,TE_B0ref,T1_GM(refB0idx),T2s_GM(refB0idx));
-DWs_GM_B0ref  = DWSignal(b,D_GM);
-S_GM_B0ref    = LongSSs_GM_B0ref * DWs_GM_B0ref;
-SNR_GM_ref    = SNR_B0ref * 1/sqrt(TR_B0ref) * 1/sqrt(BW_B0ref) * S_GM_B0ref * (resNEW(refRes)/resINI)^3;
+TransvSSs_GM_B0ref = TransvSS_Signal(OptimumAlpha_GM_B0ref,TR_B0ref,TE_B0ref,T1_GM(refB0idx),T2s_GM(refB0idx));
+DWs_GM_B0ref       = DWSignal(b,D_GM);
+S_GM_B0ref         = TransvSSs_GM_B0ref * DWs_GM_B0ref;
+SNR_GM_ref         = SNR_B0ref * 1/sqrt(TR_B0ref) * 1/sqrt(BW_B0ref) * S_GM_B0ref * (resNEW(refRes)/resINI)^3;
 
 % ... WM tissue ...
-LongSSs_WM_B0ref = LongSS_Signal(OptimumAlpha_WM_B0ref,TR_B0ref,TE_B0ref,T1_WM(refB0idx),T2s_WM(refB0idx));
-DWs_WM_B0ref  = DWSignal(b,D_WM);
-S_WM_B0ref    = LongSSs_WM_B0ref * DWs_WM_B0ref;
-SNR_WM_ref    = SNR_B0ref * 1/sqrt(TR_B0ref) * 1/sqrt(BW_B0ref) * S_WM_B0ref * (resNEW(refRes)/resINI)^3;
+TransvSSs_WM_B0ref = TransvSS_Signal(OptimumAlpha_WM_B0ref,TR_B0ref,TE_B0ref,T1_WM(refB0idx),T2s_WM(refB0idx));
+DWs_WM_B0ref       = DWSignal(b,D_WM);
+S_WM_B0ref         = TransvSSs_WM_B0ref * DWs_WM_B0ref;
+SNR_WM_ref         = SNR_B0ref * 1/sqrt(TR_B0ref) * 1/sqrt(BW_B0ref) * S_WM_B0ref * (resNEW(refRes)/resINI)^3;
 
 % ... LS tissue ...
-LongSSs_LS_B0ref = LongSS_Signal(OptimumAlpha_LS_B0ref,TR_B0ref,TE_B0ref,T1_LS(refB0idx),T2s_LS(refB0idx));
-DWs_LS_B0ref  = DWSignal(b,D_LS);
[…]
