### 2020-08-05 #2478 @Kevin-Mattheus-Moerman
@ritagnunes can you provide an update on progress in relation to responding to the issues raised?
