### 2020-08-05 #2478 @ritagnunes
We are working towards solving these issues asap.

Regarding the contribution of each author, here is the answer to
@mathieuboudreau <https://github.com/mathieuboudreau> 's question:

@ritagnunes - joint implementation of write_*py sequence files,
implementation and tests for some of the auxiliary diffusion functions,
implementation and testing of basic reconstruction functions, joint project
supervision and manuscript preparation
@sravan953 - performing MRI experiments for testing the generated sequences
and generating example data, providing basic documentation and testing code
@opennog - joint implementation of write_*.py sequence files,
implementation and tests for some of the auxiliary diffusion functions
@ZemaTimoteo - implementation of the SNR study code, its auxiliary
functions and corresponding tests
@mvgolub - theoretical derivation of b-value optimization for SNR study,
review and testing of the SNR study code
@PavanPoojar -  provided basic documentation and testing code
@sairamgeethanath - provided basic documentation, joint project supervision
and manuscript preparation

On Wed, Aug 5, 2020 at 11:28 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @ritagnunes <https://github.com/ritagnunes> can you provide an update on
> progress in relation to responding to the issues raised?
>
