### 2020-08-05 8ae7611
Update README.md
Provided more explicit usage instructions.
