+++ README.md
@@ -8,7 +8,7 @@
 
 Diffusion-weighted Imaging (DWI) is an important sequence for many clinical applications such as stroke and tumor characterization [[1]](#references).  Though several vendor-neutral post-processing tools are available for diffusion, open-source acquisition implementations have not been provided for research purposes.
 
-We propose the development of a tool `PulseqDiffusion` cross-vendor, open-source package of a multi-slice single-shot spin-echo echo-planar imaging (EPI)-based diffusion pulse sequence using `PyPulseq` [[2]](#references) - [[4]](#references), which can be extended to support multiple b-values and directions. We demonstrate this on (i) in vitro phantom to measure the Apparent Diffusion Coefficient  (ii) in vivo human brain data to obtain good quality Fractional Anisotropy (FA) and Mean Diffusivity maps. We provide basic reconstrucion software implemented in Matlab, for generating the images from the raw k-space data (main Matlab script: process_data.m). Example k-space data acquired at the University of Columbia is provided in the folder Example_Data (Phantom data: 3 directions, 5 b-values, 3 slices; In-Vivo data: 12 directions, 1 b-value, 20 slices and 3 directions, 3 b-values, 20 slices).
+We propose the development of a tool `PulseqDiffusion` cross-vendor, open-source package of a multi-slice single-shot spin-echo echo-planar imaging (EPI)-based diffusion pulse sequence using `PyPulseq` [[2]](#references) - [[4]](#references), which can be extended to support multiple b-values and directions. We demonstrate this on (i) in vitro phantom to measure the Apparent Diffusion Coefficient  (ii) in vivo human brain data to obtain good quality Fractional Anisotropy (FA) and Mean Diffusivity (MD) maps. We provide basic reconstrucion software implemented in Matlab, for generating the images from the raw k-space data. Example data acquired at the University of Columbia is provided in the folder `Example_Data` (Phantom data: 3 directions, 5 b-values, 3 slices; In-Vivo data: 12 directions, 1 b-value, 20 slices and 3 directions, 3 b-values, 20 slices).
 
 The images can then be processed utilizing freely available post-processing tools for generating quantitative diffusion maps. Example code using the image analysis software FSL (https://fsl.fmrib.ox.ac.uk/fsl/fslwiki) [[5]](#references) is included in this package.
 
@@ -29,6 +29,28 @@ https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FslInstallation
 
 Note: raw MRI data readers are vendor specific and hence cannot be provided.
 
+## Usage and Examples
+1- The user should start by running one of the three Python scripts provided to generate a diffusion-weighted EPI sequence:
+`write_se_dw_epi_bValue.py`, `write_se_dw_rsepi_bValue.py`, `write_trse_dw_epi_bValue.py` 
+The first two use single spin-echo preparation modules, while in the latter a doubly-refocused diffusion preparation module is provided for reducing eddy-current distortions.
+
+Each of these scripts contains a section called "Acquisition Parameters" which can be modified as desired (e.g. k-space matrix size - Nx, Ny; slice_thickness).
+Importantly, the system limits should be set according to the used MRI scanner.
+Basic diffusion schemes are provided in the auxiliary file `diff_funcs.py` (3, 6, 12 and 60 directions), which can easily be extended by the user.
+
[…]
