### 2020-08-05 ad5b083
Update README.md
Included information on how to generate API documentation.
