+++ README.md
@@ -10,7 +10,7 @@ Diffusion-weighted Imaging (DWI) is an important sequence for many clinical appl
 
 We propose the development of a tool `PulseqDiffusion` cross-vendor, open-source package of a multi-slice single-shot spin-echo echo-planar imaging (EPI)-based diffusion pulse sequence using `PyPulseq` [[2]](#references) - [[4]](#references), which can be extended to support multiple b-values and directions. We demonstrate this on (i) in vitro phantom to measure the Apparent Diffusion Coefficient  (ii) in vivo human brain data to obtain good quality Fractional Anisotropy (FA) and Mean Diffusivity (MD) maps. We provide basic reconstrucion software implemented in Matlab, for generating the images from the raw k-space data. Example data acquired at the University of Columbia is provided in the folder `Example_Data` (Phantom data: 3 directions, 5 b-values, 3 slices; In-Vivo data: 12 directions, 1 b-value, 20 slices and 3 directions, 3 b-values, 20 slices).
 
-The images can then be processed utilizing freely available post-processing tools for generating quantitative diffusion maps. Example code using the image analysis software FSL (https://fsl.fmrib.ox.ac.uk/fsl/fslwiki) [[5]](#references) is included in this package.
+The images can then be processed utilizing freely available post-processing tools for generating quantitative diffusion maps. Example code using the image analysis software [FSL](https://fsl.fmrib.ox.ac.uk/fsl/fslwiki) [[5]](#references) is included in this package.
 
 A tool for predicting the signal-to-noise ratio (SNR) observed for different brain tissues (gray matter, white matter, cerebrospinal fluid - CSF) and the contrast-to-noise ratio (CNR) of an acute stroke lesion relative to those tissues can be found in `SNR_CNR_Study`. The implemented code predicts the SNR and CNR per time unit for a spin-echo diffusion-weighted sequence, taking into account the steady-state value for the longitudinal magnetization. For that purpose, we adapted an expression used to predict the SNR per tissue per time unit for a spoiled gradient echo sequence [[6]](#references). This simulation considers the possibility to use either EPI or spiral readouts and considers the impact of: B0, max gradient amplitude, spatial resolution and b-value (comparing the achieved SNR with that of a typical 1.5T clinical scanner).
 
@@ -23,7 +23,7 @@ For sequence development: PyPulseq [[2]](#references) \>=Python 3.6, virtual env
 To perform basic image reconstruction and processing:
 Matlab toolbox for writing and reading Nifti images: Nifti_tools downloadable from [[7]](#references) to be placed within a folder called 'nifti_tools'
 
-FSL - Fmrib's Software Library [[5]](#references)
+[FSL](https://fsl.fmrib.ox.ac.uk/fsl/fslwiki) - Fmrib's Software Library [[5]](#references)
 FSL installation instructions can be found at:
 https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FslInstallation
 
@@ -44,12 +44,20 @@ A .seq file will be written which can be run in different MRI platforms. In orde
 By specifying the name of the k-space data file provided in the [Example_Data](https://github.com/ritagnunes/PulseqDiffusion/tree/master/Example_Data) folder, the user will be able to reconstruct the corresponding images and estimate either the full diffusion tensor (using FSL's ´dtifit´) or directional apparent diffusion coefficient maps (by setting the "model" parameter). Note that this script will need to be consistent with the selected acquisition parameters (specified spatial resolution and slice thickness).
 The script enables to use the FSL "eddy" pre-processing tool by setting a parameter with the same name to 1.
 
[…]
