### 2020-08-05 ritagnunes/PulseqDiffusion#6 @ritagnunes
Documentation can be generated by compiling the Sphinx documents provided under the docs folder.
I have edited the README.md file to explicitly explain how this can be achieved (commit ad5b083).
