### 2020-08-05 ritagnunes/PulseqDiffusion#4 @ritagnunes
Yes, hopefully the Usage instructions that have been added to README.md in a recent commit will help to make this clearer.
