### 2020-08-05 fd4ac3e
Update CONTRIBUTING.md
Response to:  [JOSS] Broken link for "Code of conduct" in the "CONTRIBUTING.MD" page #2
