### 2020-08-06 99d32ac
Create assertValue.m
