### 2020-08-06 f625e8f
Update recon_imgsTest.m
moved assertValue function to separate file
