+++ PulseqDiffusion/recon_imgsTest.m
@@ -18,11 +18,3 @@ tmp=load('../Example_Data/ImgPhantom3dirs5b3slices.mat');
 impred=tmp.im_sos;
 imact=recon_imgs(t.k, t.dim_struct);
 assertValue(imact,impred)
-
-function assertValue(actVal,expVal)
-% Helper function to assert equality
-% Takes two images and compares them 
-
-tf = isequal(actVal,expVal);
-assert(tf, 'Recon images do not match!');
-end
