### 2020-08-06 55a5fe3
Update generate_eddyfilesTest.m
moved assertFiles function to a separate file
