### 2020-08-06 5322738
Create assertFiles.m
created assertFiles auxiliary function for testing
