### 2020-08-06 12c2391
Update generate_bvalsbvecsTest.m
Moved assertWithAbsTol helper function to a separate file
