### 2020-08-06 c77eb54
Create assertWithAbsTol.m
Created separate file assertWithAbsTol.m for auxiliary test function
