+++ PulseqDiffusion/assertWithAbsTol.m
@@ -1,0 +1,8 @@
+function assertWithAbsTol(actVal,expVal,varargin)
+% Helper function to assert equality within an absolute tolerance.
+% Takes two values and an optional message and compares
+% them within an absolute tolerance of 1e-6.
+tol = 1e-6;
+tf = norm(actVal-expVal) <= tol;
+assert(tf, varargin{:});
+end  
