### 2020-08-06 2c5bcb9
Update estimate_adcTest.m
Moved assertValue function to a separate file.
