### 2020-08-07 cee2c38
Allowed for numerical tolerance when testing matlab scripts
