### 2020-08-07 ritagnunes/PulseqDiffusion#4 @grlee77
Yes, it seems this is not a traditional software library with an installation process, but is instead a series of Python and Matlab scripts.

The Python scripts generate the sequence files that would be used on the scanner and the Matlab scripts are for reconstruction, generating gradient tables, etc.

The Python code can be run as long as the user has installed `pypulseq` as indicated in the README. Then there is no actual installation step for PulseqDiff itself, one just navigates to the folder with the scripts and can run them directly from the shell via
`python write_se_dwi_epi_bValue.py`, etc.
