### 2020-08-07 #2478 @grlee77
See comment https://github.com/ritagnunes/PulseqDiffusion/issues/4#issuecomment-670660507 regarding how this package is a mixture of python and Matlab scripts and not a traditional, installable software library . Given, that, why did the team decide to use the Python-based pypulseq, but keep the other scripts in Matlab?

For example, `recon_img.m` looks like it would be easy to convert from Matlab to Python using just functions in the NumPy library (fft, fftshift, squeeze, etc.). If the authors prefer Matlab, then why use pypulseq rather than the original [Matlab-based pulseq](https://pulseq.github.io/matlab.html) for the pulse sequence generation scripts? I personally prefer Python, but realize Matlab is also very popular in MR research labs. It just seems a little odd to mix the two rather than choosing one or the other.

Also, it looks like these Matlab scripts don't rely on specific toolboxes and probably just work in Octave? If that is the case, I would mention that so that Python users who might be interested will know that they don't necessarily need a commercial Matlab license to run the scripts.
