### 2020-08-08 675d8b9
modified assert test fucntion
