### 2020-08-09 ritagnunes/PulseqDiffusion#4 @ritagnunes
At present we are using Matlab for reconstructing the images as this code spans from previous work, developed over time.
In any case we would like to point out that as these scripts don't rely on specific toolboxes, users who do not have access to Matlab should be able to use Octave instead. We are planning to provide a Python version of the reconstruction code in a future release.
