### 2020-08-20 #2478 @Kevin-Mattheus-Moerman
@mathieuboudreau, @grlee77 if you attended ISMRM I hope you enjoyed it. This is just a friendly check in to see where we stand with this submission. Let me know if you have any questions.
