### 2020-08-24 #2478 @mathieuboudreau
Thanks to some clarifications from the authors (https://github.com/ritagnunes/PulseqDiffusion/issues/4#issuecomment-669341558, https://github.com/ritagnunes/PulseqDiffusion/issues/6), I’ve been able to better understand the authors goal for the code and test it out a bit more.

After testing the code a bit, to my understanding (and as @grlee77 pointed out) this submission is more of a collection of scripts/functions than a piece of software. There are two main problems that this code tries to be a solution to from my understanding : #1 create a diffusion pulse sequence (e.g. list of events to be executed by the MRI hardware) with the PyPulseq tool (authors provide three demos), and #2 take raw data that’s acquired from the MRI hardware using the above pulse sequence(s) and reconstruct images that can be used in analyses.

The authors implementation for #1 are just Python scripts (wrapped in a `__main__` function) where all the relevant values are hardcoded (i.e. can't be passed to the function), and the user is asked without real guidance to modify or update them inside the scripts for their own particular application. The scripts depend entirely on another piece of software (PyPulseq), and in themselves could have been just as easily integrated as addons to that software instead of this proposed standalone PulseqDiffusion. I checked, and there is also substantial duplicate code between two of the three scripts offered by the authors (write_se_dwi_epi_bValue.py and write_se_dwi_rsepi_bValue.py), thus these scripts could in fact be refactored into modular functions that could then be tested and a single API could be offered to the user. Also, hardcoding parameters that need to be changed by the user inside the files increases the level of difficulty in terms of user experience (providing a way to pass along these parameters as a file would be better in my view, as then these files could be shared or kept track of).

The authors implementation for #2 are MATLAB files that are completely separate from the implementation for #1. I tried running their example again today, but was unable to run it as it appears that my fresh installation of FSL (a dependency) doesn’t have a file called by the authors (missing eddy_openmp, though my installation has eddy eddy_correct eddy_quad eddy_combine  eddy_cuda7.5 eddy_squada). It may just be that the authors have an older version of FSL installed, although no specific version for any of their dependencies are stated in their installation instructions.

I think overall, due to the presence of a lack of coherence in the code (no link between parts 1 and 2 above in terms of code or user experience) and strong dependency on other softwares (part 1 is essentially scripts implementing PyPulseq use cases, but isn’t a piece of software in itself), I’m really on the fence about if this merits a tick on the JOSS checkbox:

> Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

In particular, this line in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) is where I have pause here:

>In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools). “Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable.

As with @grlee77, I think this work has value and that substantial work has been done by the authors, and by looking at the code I don’t have doubts on the quality or MR physics of their implementation. But in my view, I don’t think in its current state it represents a coherent piece of software (i.e. "feature complete"?), but rather a collection of scripts & tools ("minor utility" package?). Does this warrant a publication in JOSS? As a somewhat newbie JOSS reviewer, I’m leaning towards no in its current state, but think with an improved user experience and more coherence between all the parts of the code, it could be done (but would require significantly more work). My mind could be changed though if the JOSS editor (@Kevin-Mattheus-Moerman) chimes in and says that what I described above is within the scope of what JOSS wants from their submissions, if so then I’d have additional comments for suggestions to improve the software (since I still haven’t been able to test it all yet without issues).
