### 2020-08-25 #2478 @Kevin-Mattheus-Moerman
Thank you @mathieuboudreau for highlighting these important points.

@grlee77 it would be great if you could also weigh in on these comments. Based on @mathieuboudreau 's comments this work does not warrant publication in JOSS (given https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). Do you agree with his statements?
