### 2020-08-26 3c040f7
Replaced norm call
