### 2020-08-26 #2478 @ritagnunes
Hi  @grlee77, I believe that is a Matlab version issue. The behaviour of
the norm function must have changed.
We have modified the code to use our own implementation. Let me know if you
still have problems.

Thanks

On Tue, Aug 25, 2020 at 3:15 PM Gregory R. Lee <notifications@github.com>
wrote:

> Do you agree with his statements?
>
> Yes, I think those comments are accurate and was also uncertain that it
> met those the criteria. I agree with his summary:
>
> But in my view, I don’t think in its current state it represents a
> coherent piece of software (i.e. "feature complete"?), but rather a
> collection of scripts & tools ("minor utility" package?).
>
> The only difference I see from what @mathieuboudreau
> <https://github.com/mathieuboudreau> reported is that the process_data.m
> example did work for me (FSL 5.0.11 on my system had the required
> eddy_openmp binary).
>
> I did also try running the unit tests today via
> runTests_ReconAndProcessing.m. Some passed, but others failed with
> complaints about Matlab's norm function expecting 2D data. I don't know
> if this is an issue of a difference in Matlab version (I tried with R2017a)
> or if the authors had a custom n-dimensional norm.m implementation that
> is missing from the repo. Details are reproduced below.
>
> Error occurred in recon_imgsTest/TestExampleInVivoData12Directions_Nb1_20Slices_1 and it did not run to completion.
>
>
>
>     ---------
>
>     Error ID:
>
>     ---------
>
>     'MATLAB:norm:inputMustBe2D'
>
>
>
>     --------------
>
>     Error Details:
>
>     --------------
>
>     Error using norm
>
>     Input must be 2-D.
>
>
>
>     Error in assertWithAbsTol (line 6)
>
>     tf = norm(actVal-expVal) <= tol;
>
>
>
>     Error in recon_imgsTest (line 20)
>
>     assertWithAbsTol(imact,impred,'Reconstructed images do not match predictions: Phantom3dirs5b3slices')
>
> ================================================================================
>
> .
>
> Done recon_imgsTest
>
> __________
>
>
>
> Failure Summary:
>
>
>
>      Name                                                             Failed  Incomplete  Reason(s)
>
>     ================================================================================================
>
>      estimate_adcTest/TestExampleInVivoData3Directions_Nb3_20Slices     X         X       Errored.
>
>     ------------------------------------------------------------------------------------------------
>
>      estimate_adcTest/TestExamplePhantomData3Directions_Nb5_3Slices     X         X       Errored.
>
>     ------------------------------------------------------------------------------------------------
>
>      recon_imgsTest/TestExampleInVivoData12Directions_Nb1_20Slices      X         X       Errored.
>
>     ------------------------------------------------------------------------------------------------
>
>      recon_imgsTest/TestExampleInVivoData3Directions_Nb3_20Slices       X         X       Errored.
>
>     ------------------------------------------------------------------------------------------------
>
>      recon_imgsTest/TestExampleInVivoData12Directions_Nb1_20Slices_1    X         X       Errored.
>
>
>
>
>
> testResults =
>
>
>
>   1×15 TestResult array with properties:
>
>
>
>     Name
>
>     Passed
>
>     Failed
>
>     Incomplete
>
>     Duration
>
>     Details
>
>
>
> Totals:
>
>    10 Passed, 5 Failed, 5 Incomplete.
>
>    17.0859 seconds testing time.
>
>
> <\details>
>
